requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
